Split det tests
